+++ packages/terra-compact-interactive-list/src/proptypes/cellShape.js
@@ -5,16 +5,6 @@ const cellShape = PropTypes.shape({
    * Content that will be rendered within the Cell.
    */
   content: PropTypes.node,
-  /**
-   * @private
-   * Boolean value indicating whether or not the column header is selectable.
-   */
-  isSelectable: PropTypes.bool,
-  /**
-   * @private
-   * Boolean value indicating whether or not the cell should render as selected.
-   */
-  isSelected: PropTypes.bool,
 });
 
 export default cellShape;