+++ packages/terra-data-grid/src/DataGrid.jsx
@@ -2,15 +2,13 @@ import React, {
   useState, useRef, useCallback, forwardRef, useImperativeHandle, useMemo,
 } from 'react';
 import PropTypes from 'prop-types';
-import { injectIntl } from 'react-intl';
 import classNames from 'classnames/bind';
 import * as KeyCode from 'keycode-js';
-import Table, { GridConstants, GridContext } from 'terra-table';
+import Table, {
+  GridConstants, GridContext, sectionShape, rowShape, columnShape, validateRowHeaderIndex,
+} from 'terra-table';
 import VisuallyHiddenText from 'terra-visually-hidden-text';
-import rowShape from './proptypes/rowShape';
-import { columnShape } from './proptypes/columnShape';
 import WorklistDataGridUtils from './utils/WorklistDataGridUtils';
-import validateRowHeaderIndex from './proptypes/validators';
 import styles from './DataGrid.module.scss';
 import './_elementPolyfill';
 
@@ -37,6 +35,11 @@ const propTypes = {
    * Data for content in the body of the Grid. Rows will be rendered in the order given.
    */
   rows: PropTypes.arrayOf(rowShape),
+
+  /**
+  * Data for content in the body of the table. Sections will be rendered in the order given.
+  */
+  sections: PropTypes.arrayOf(sectionShape),
 
   /**
    * Data for pinned columns. Pinned columns are the stickied leftmost columns of the grid.
@@ -88,6 +91,11 @@ const propTypes = {
    * @param {string} columnId columnId
    */
   onCellSelect: PropTypes.func,
+
+  /**
+   * Function that is called when a collapsible section is selected. Parameters: `onSectionSelect(sectionId)`
+   */
+  onSectionSelect: PropTypes.func,
 
   /**
    * Callback function that is called when a selectable column is selected. Parameters:
@@ -144,7 +152,7 @@ const defaultProps = {
   hasVisibleColumnHeaders: true,
 };
 
-const DataGrid = injectIntl((props) => {
+const DataGrid = forwardRef((props, ref) => {
   const {
     ariaLabel,
     ariaLabelledBy,
@@ -161,11 +169,13 @@ const DataGrid = injectIntl((props) => {
     onColumnSelect,
     onRangeSelection,
     onRowSelectionHeaderSelect,
+    onSectionSelect,
     overflowColumns,
     pinnedColumns,
     rowHeaderIndex,
     rowHeight,
     rows,
+    sections,
   } = props;
 
   const displayedColumns = (hasSelectableRows ? [WorklistDataGridUtils.ROW_SELECTION_COLUMN] : []).concat(pinnedColumns).concat(overflowColumns);
@@ -184,6 +194,8 @@ const DataGrid = injectIntl((props) => {
   const gridContainerRef = useRef();
   const tableContainerRef = useRef();
   const handleFocus = useRef(true);
+
+  const focusedCellRef = useRef({ rowId: '', columnId: '' });
 
   const [checkResizable, setCheckResizable] = useState(false);
 
@@ -205,14 +217,19 @@ const DataGrid = injectIntl((props) => {
   // -------------------------------------
   // functions
 
-  const isRowSelectionCell = (columnIndex) => (
+  const isRowSelectionCell = useCallback((columnIndex) => (
     hasSelectableRows && columnIndex < displayedColumns.length && displayedColumns[columnIndex].id === WorklistDataGridUtils.ROW_SELECTION_COLUMN.id
-  );
+  ), [displayedColumns, hasSelectableRows]);
 
-  const setFocusedRowCol = (newRowIndex, newColIndex, makeActiveElement) => {
+  const setFocusedRowCol = useCallback((newRowIndex, newColIndex, makeActiveElement) => {
     setCellAriaLiveMessage(null);
     setFocusedRow(newRowIndex);
     setFocusedCol(newColIndex);
+
+    focusedCellRef.current = {
+      rowId: grid.current.rows[newRowIndex].getAttribute('data-row-id'),
+      columnId: displayedColumns[newColIndex].id,
+    };
 
     if (makeActiveElement) {
       // Set focus on input field (checkbox) of row selection cells.
@@ -228,14 +245,14 @@ const DataGrid = injectIntl((props) => {
 
       focusedCell?.focus();
     }
-  };
+  }, [isRowSelectionCell, displayedColumns]);
 
   // The focus is handled by the DataGrid. However, there are times
   // when the other components may want to change the currently focus
   // cells. In order to do so, these datagrid methods will be exposed to
   // allow those components to request focus change.
   useImperativeHandle(
-    props.focusFuncRef,
+    ref,
     () => ({
       setFocusedRowCol,
       getFocusedCell() { return { row: focusedRow, col: focusedCol }; },
@@ -285,28 +302,27 @@ const DataGrid = injectIntl((props) => {
 
   const handleColumnSelect = useCallback((columnId) => {
     const columnIndex = displayedColumns.findIndex(column => column.id === columnId);
-    setFocusedCol(columnIndex);
+    setFocusedRowCol(0, columnIndex);
 
     if (onColumnSelect) {
       onColumnSelect(columnId);
     }
-  }, [onColumnSelect, displayedColumns]);
+  }, [onColumnSelect, displayedColumns, setFocusedRowCol]);
 
   const handleRowSelectionHeaderSelect = useCallback(() => {
-    setFocusedCol(0);
-    setFocusedRow(0);
+    setFocusedRowCol(0, 0);
     if (onRowSelectionHeaderSelect) {
       onRowSelectionHeaderSelect();
     }
-  }, [onRowSelectionHeaderSelect]);
+  }, [onRowSelectionHeaderSelect, setFocusedRowCol]);
 
   const handleCellSelection = useCallback((selectionDetails) => {
-    setFocusedRow(selectionDetails.rowIndex);
-    setFocusedCol(selectionDetails.columnIndex);
+    const { columnIndex, rowIndex } = selectionDetails;
+    setFocusedRowCol(rowIndex, columnIndex);
     if (onCellSelect) {
       onCellSelect(selectionDetails);
     }
-  }, [onCellSelect]);
+  }, [onCellSelect, setFocusedRowCol]);
 
   // -------------------------------------
   // event handlers
@@ -468,6 +484,7 @@ const DataGrid = injectIntl((props) => {
       event.preventDefault(); // prevent the page from moving with the arrow keys.
       return;
     }
+
     handleMoveCellFocus(cellCoordinates, { row: nextRow, col: nextCol });
     event.preventDefault(); // prevent the page from moving with the arrow keys.
   };
@@ -485,7 +502,26 @@ const DataGrid = injectIntl((props) => {
     if (!event.currentTarget.contains(event.relatedTarget)) {
       // Not triggered when swapping focus between children
       if (handleFocus.current) {
-        setFocusedRowCol(focusedRow, focusedCol, true);
+        let newRowIndex = focusedRow;
+        let newColumnIndex = focusedCol;
+
+        // Check for last focused row ID. If found set the index. Otherwise set it to the last focused row or last index.
+        if (focusedCellRef.current.rowId) {
+          newRowIndex = [...grid.current.rows].findIndex(row => row.getAttribute('data-row-id') === focusedCellRef.current.rowId);
+          newRowIndex = newRowIndex === -1
+            ? Math.min(focusedRow, grid.current.rows.length - 1)
+            : newRowIndex;
+        }
+
+        // Check for last focused column ID. If found set the index. Otherwise set it to the last focused column or last index.
+        if (focusedCellRef.current.columnId) {
+          newColumnIndex = displayedColumns.findIndex(column => column.id === focusedCellRef.current.columnId);
+          newColumnIndex = newColumnIndex === -1
+            ? Math.min(focusedCol, displayedColumns.length - 1)
+            : newColumnIndex;
+        }
+
+        setFocusedRowCol(newRowIndex, newColumnIndex, true);
         setGridHasFocus(true);
       }
     }
@@ -518,6 +554,7 @@ const DataGrid = injectIntl((props) => {
         <Table
           id={`${id}-table`}
           rows={dataGridRows}
+          sections={sections}
           ariaLabelledBy={ariaLabelledBy}
           ariaLabel={ariaLabel}
           activeColumnIndex={(gridHasFocus && focusedRow === 0) ? focusedCol : undefined}
@@ -531,6 +568,7 @@ const DataGrid = injectIntl((props) => {
           rowHeaderIndex={rowHeaderIndex}
           onColumnResize={onColumnResize}
           onColumnSelect={handleColumnSelect}
+          onSectionSelect={onSectionSelect}
           onCellSelect={handleCellSelection}
           onRowSelectionHeaderSelect={handleRowSelectionHeaderSelect}
           hasSelectableRows={hasSelectableRows}
@@ -543,7 +581,7 @@ const DataGrid = injectIntl((props) => {
   );
 });
 
-DataGrid.propTypes = propTypes;
 DataGrid.defaultProps = defaultProps;
+DataGrid.propTypes = propTypes;
 
-export default forwardRef((props, ref) => <DataGrid {...props} focusFuncRef={ref} />);
+export default DataGrid;