+++ packages/terra-table/tests/jest/ColumnHeader.test.jsx
@@ -26,6 +26,8 @@ describe('ColumnHeader', () => {
     const columnHeader = wrapper.find('.column-header-row');
     expect(columnHeader).toHaveLength(1);
     expect(columnHeader.props().height).toBe('3rem');
+    expect(columnHeader.props()['data-row-id']).toBe('test-table-header-row');
+    expect(columnHeader.props()['aria-rowindex']).toBe(1);
 
     //  Validate ColumnHeaderCell React component
     const columnHeaderCells = columnHeader.find(ColumnHeaderCell);
@@ -57,6 +59,8 @@ describe('ColumnHeader', () => {
     const columnHeader = wrapper.find('.column-header-row');
     expect(columnHeader).toHaveLength(1);
     expect(columnHeader.props().height).toBe('3rem');
+    expect(columnHeader.props()['data-row-id']).toBe('test-table-header-row');
+    expect(columnHeader.props()['aria-rowindex']).toBe(1);
 
     // Validate ColumnHeaderCell React component
     const columnHeaderCells = columnHeader.find(ColumnHeaderCell);
@@ -87,6 +91,8 @@ describe('ColumnHeader', () => {
     const columnHeader = wrapper.find('.column-header-row');
     expect(columnHeader).toHaveLength(1);
     expect(columnHeader.props().height).toBe('3rem');
+    expect(columnHeader.props()['data-row-id']).toBe('test-table-header-row');
+    expect(columnHeader.props()['aria-rowindex']).toBe(1);
 
     // Validate ColumnHeaderCell React component
     const columnHeaderCells = columnHeader.find(ColumnHeaderCell);
@@ -123,6 +129,8 @@ describe('ColumnHeader', () => {
     const columnHeader = wrapper.find('.column-header-row');
     expect(columnHeader).toHaveLength(1);
     expect(columnHeader.props().height).toBe('3rem');
+    expect(columnHeader.props()['data-row-id']).toBe('test-table-header-row');
+    expect(columnHeader.props()['aria-rowindex']).toBe(1);
 
     // Validate ColumnHeaderCell React component
     const columnHeaderCell = columnHeader.find(ColumnHeaderCell).first();
@@ -155,6 +163,8 @@ describe('ColumnHeader', () => {
     // Verify that column headers are not present
     const columnHeader = wrapper.find('.hidden');
     expect(columnHeader).toHaveLength(1);
+    expect(columnHeader.props()['data-row-id']).toBe('test-table-header-row');
+    expect(columnHeader.props()['aria-rowindex']).toBe(1);
 
     expect(wrapper).toMatchSnapshot();
   });